[DPP-4937] Advertising and Outlook Fixes
There were some mismatched conditional comment tags for Outlook-specific code that were broken. This caused some layout bugs, mostly to the display ads, on Outlook that an advertiser noticed. I traced the problem to extra closing table tags on the advertisement block. The start of these was lost in one of the iterations but has been restored in the emailx.marko file, along with moving the closing tags here to simplify the code. There was also an extraneous set of closing table tags in the content-hero element that have been removed.