[DPP-4916] Hero Image Optimizations
The image options were recently changed to the incorrect params ("width" and "height" instead of w and h) which was causing the hero images not to crop. This resulted in some rather large (long) images displaying in the newsletters and at a default resolution of 320. This fix should set them back to the intended 600px x 300px. During my testing I happen to have a really bad image that didn't crop well once the fix was applied. But I added the facial crop property to improve this. This should solve a lot of display issues we saw with the hero images (GCI has a few 'good' examples)