[DPP-5156] Updated structure for digital edition emails
standardizing the structure, since they varied a bit.
Adding a better spacer in-between the cover/ad row and the bottom promos.
turning off in-this-issue-block and turning on digital-edition-overview-block above the cover/ad row.
Adding the new digital-edition-highlight-block to support Promotions in the second (and third) position to use the Featured Hero layout.