[DPP-5156] Using isDigital to turn off DPM on Promos
Promotions default as 'Advertisements' in the DPM code, but shouldn't be. The proposed solution was to turn off DPM for any of the links in the bottom 'highlight' promotions. To do this, I am utilizing the isDigital flag previously set up and passing that into the dpm object. And on content dpm items, we use this flag to disable dpm tracking. This approach allows us to change how these are coded, if they change their mind, without having to change how the dpm code works for all of the other newsletter types.