Digital Mag Email items will link to Digital Mag
This change will allow Editors to schedule normal content items to the Digital Edition Email Newsletters and have them displayed as highlighted items, but link directly to the digital edition link instead of their website urls. This will allow them to not have to create so many near-duplicate Promotion items for each one they want to highlight in the newsletters.